Umbraco deprecated the MediaService.Created event in 6.2x so we had to move the creation logic to the Saving event
I also did the same thing in ArticleEvents assuming that Created will be deprecated there at some point as well

+++ PPC 2010/UmbracoEvents/SermonEvents.cs
@@ -15,7 +15,6 @@ namespace PPC_2010.UmbracoEvents
     {
         public void OnApplicationInitialized(UmbracoApplicationBase umbracoApplication, ApplicationContext applicationContext)
         {
-            MediaService.Created += new TypedEventHandler<IMediaService, NewEventArgs<IMedia>>(MediaService_Created);
             MediaService.Saving += new TypedEventHandler<IMediaService, SaveEventArgs<IMedia>>(MediaService_Saving);
             MediaService.Saved += new TypedEventHandler<IMediaService, SaveEventArgs<IMedia>>(MediaService_Saved);
             MediaService.Deleted += new TypedEventHandler<IMediaService, DeleteEventArgs<IMedia>>(MediaService_Deleted);
@@ -26,57 +25,43 @@ namespace PPC_2010.UmbracoEvents
 
         public void OnApplicationStarting(UmbracoApplicationBase umbracoApplication, ApplicationContext applicationContext) { }
 
-        void MediaService_Created(IMediaService sender, NewEventArgs<IMedia> e)
-        {
-            if (e.Entity.ContentType.Alias == PPC_2010.Data.Constants.SermonAlias)
-            {
-                // Automatically populate the sermon title from what the user entered in for the element name
-                // The element name will be set in Media_BeforeSave and having the name they type in when
-                // creating a new sermon makes more sense to users
-                var sermon = new Data.Media.MediaSermon(e.Entity);
-                sermon.Title = e.Entity.Name;
-                sermon.RecordingDate = DateTime.Today;
-
-                CheckForRefresh(e.Entity, e);
-            }
-        }
-
         void MediaService_Saving(IMediaService sender, SaveEventArgs<IMedia> e)
         {
             foreach (IMedia entity in e.SavedEntities)
             {
-                if (entity.ContentType.Alias == PPC_2010.Data.Constants.SermonAlias && entity.IsValid())
+                if (entity.ContentType.Alias == PPC_2010.Data.Constants.SermonAlias)
                 {
                     if (CheckForRefresh(entity, e))
                         return;
 
-                    Data.Media.MediaSermon sermon = new Data.Media.MediaSermon(entity);
-
-                    // don't update anything recorded before the launch of the new web site
-                    if (sermon.RecordingDate > new DateTime(2011, 2, 28))
+                    var sermon = new Data.Media.MediaSermon(entity);
+                    if (!entity.HasIdentity)
+                    {
+                        // Automatically populate the sermon title from what the user entered in for the element name
+                        // The element name will be set in Media_BeforeSave and having the name they type in when
+                        // creating a new sermon makes more sense to users
+                        sermon.Title = entity.Name;
+                        sermon.RecordingDate = DateTime.Today;
+                    }
+                    else if (entity.IsValid() && sermon.RecordingDate > new DateTime(2011, 2, 28)) // don't update anything recorded before the launch of the new web site
                     {
                         // Canonicalize the name of the sermon so they are consistent and easy to find in the list
                         entity.Name = sermon.RecordingDate.Value.ToString("MM/dd/yyyy") + " - " + sermon.RecordingSession;
 
-                        // Cananicalize the file name,
-                        if (sermon.RecordingDate > new DateTime(2011, 2, 28))
+                        string oldFileName = entity.GetValue<string>("audioFile"); ;
+                        if (!string.IsNullOrWhiteSpace(oldFileName) && System.IO.File.Exists(HttpContext.Current.Server.MapPath(oldFileName)))
                         {
-                            string oldFileName = entity.GetValue<string>("audioFile"); ;
+                            string newFileName = Path.GetDirectoryName(oldFileName) + "/" + entity.Name.Replace("/", "-") + Path.GetExtension(oldFileName);
+                            newFileName = newFileName.Replace("\\", "/").Replace(" ", "");
 
-                            if (!string.IsNullOrWhiteSpace(oldFileName) && System.IO.File.Exists(HttpContext.Current.Server.MapPath(oldFileName)))
+                            if (oldFileName != newFileName)
                             {
-                                string newFileName = Path.GetDirectoryName(oldFileName) + "/" + entity.Name.Replace("/", "-") + Path.GetExtension(oldFileName);
-                                newFileName = newFileName.Replace("\\", "/").Replace(" ", "");
+                                entity.SetValue("audioFile", newFileName);
 
-                                if (oldFileName != newFileName)
-                                {
-                                    entity.SetValue("audioFile", newFileName);
+                                if (System.IO.File.Exists(HttpContext.Current.Server.MapPath(newFileName)))
+                                    System.IO.File.Delete(HttpContext.Current.Server.MapPath(newFileName));
 
-                                    if (System.IO.File.Exists(HttpContext.Current.Server.MapPath(newFileName)))
-                                        System.IO.File.Delete(HttpContext.Current.Server.MapPath(newFileName));
-
-                                    System.IO.File.Move(HttpContext.Current.Server.MapPath(oldFileName), HttpContext.Current.Server.MapPath(newFileName));
-                                }
+                                System.IO.File.Move(HttpContext.Current.Server.MapPath(oldFileName), HttpContext.Current.Server.MapPath(newFileName));
                             }
                         }
                     }
